Add scale lock controls for flowmap legends
