+++ examples/react-app/src/App.tsx
@@ -29,13 +29,13 @@ import {
   Map as ReactMapGl,
   ViewState as ViewportProps,
 } from 'react-map-gl/maplibre';
+import {ScaleLegend} from './legends/ScaleLegend';
+import {ScaleLockButton} from './legends/ScaleLockButton';
 
 const MAP_STYLE_LIGHT =
   'https://basemaps.cartocdn.com/gl/positron-gl-style/style.json';
 const MAP_STYLE_DARK =
   'https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json';
-const LEGEND_CIRCLE_RADIUS = 10;
-const LEGEND_CIRCLE_SECONDARY_RADIUS = 7;
 
 type TooltipState = {
   position: {left: number; top: number};
@@ -160,200 +160,6 @@ function App() {
       )}
     </div>
   );
-}
-
-function ScaleLegendRoot({children}: {children: ReactNode}) {
-  return <div className="scale-legend">{children}</div>;
-}
-
-function ScaleLockButton({
-  locked,
-  onToggle,
-}: {
-  locked: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <button
-      className="scale-legend-lock-button"
-      type="button"
-      aria-pressed={locked}
-      onClick={onToggle}
-    >
-      <span className="scale-legend-lock-icon" aria-hidden="true" />
-      <span>{locked ? 'Unlock scales' : 'Lock scales'}</span>
-    </button>
-  );
-}
-
-function FlowThicknessLegend({legend}: {legend: ScaleLegendModel | undefined}) {
-  const flowThickness = legend?.flowThickness;
-  if (!flowThickness) return null;
-  return (
-    <div className="scale-legend-section scale-legend-flow-section">
-      <div className="scale-legend-title">
-        Flow thickness {legend?.locked ? '(locked)' : ''}
-      </div>
-      {flowThickness.samples.map((sample) => (
-        <div className="scale-legend-row" key={sample.label}>
-          <span
-            className="scale-legend-line"
-            style={{
-              backgroundColor: rgbaToCss(sample.color),
-              height: `${Math.max(2, sample.thickness)}px`,
-            }}
-          />
-          <span>{sample.label}</span>
-        </div>
-      ))}
-      {flowThickness.outOfScale && (
-        <div className="scale-legend-row">
-          <span
-            className="scale-legend-line"
-            style={{
-              backgroundColor: rgbaToCss(flowThickness.outOfScale.color),
-              height: `${Math.max(2, flowThickness.outOfScale.thickness)}px`,
-            }}
-          />
-          <span>{flowThickness.outOfScale.magnitudeLabel}</span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function CircleSizeLegend({legend}: {legend: ScaleLegendModel | undefined}) {
-  const locationCircles = legend?.locationCircles;
-  if (!locationCircles) return null;
-  return (
-    <div className="scale-legend-section scale-legend-circle-section">
-      <div className="scale-legend-title">
-        Circle size {legend?.locked ? '(locked)' : ''}
-      </div>
-      <CircleLegendExample
-        label="Incoming > outgoing"
-        innerRadius={LEGEND_CIRCLE_RADIUS}
-        outerRadius={LEGEND_CIRCLE_SECONDARY_RADIUS}
-        colors={locationCircles.colors}
-      />
-      <CircleLegendExample
-        label="Outgoing > incoming"
-        innerRadius={LEGEND_CIRCLE_SECONDARY_RADIUS}
-        outerRadius={LEGEND_CIRCLE_RADIUS}
-        colors={locationCircles.colors}
-      />
-      <div className="scale-legend-note">
-        Range: {formatLegendRange(locationCircles.domain)}
-      </div>
-      {locationCircles.outOfScale && (
-        <div className="scale-legend-row">
-          <span
-            className="scale-legend-circle"
-            style={{
-              backgroundColor: rgbaToCss(locationCircles.outOfScale.color),
-              width: `${LEGEND_CIRCLE_RADIUS * 2}px`,
-              height: `${LEGEND_CIRCLE_RADIUS * 2}px`,
-            }}
-          />
-          <span>{locationCircles.outOfScale.magnitudeLabel}</span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-const ScaleLegend = Object.assign(ScaleLegendRoot, {
-  FlowThickness: FlowThicknessLegend,
-  CircleSize: CircleSizeLegend,
-});
-
-function CircleLegendExample({
-  label,
-  innerRadius,
-  outerRadius,
-  colors,
-}: {
-  label: string;
-  innerRadius: number;
-  outerRadius: number;
-  colors: NonNullable<ScaleLegendModel['locationCircles']>['colors'];
-}) {
-  const size = Math.max(innerRadius, outerRadius) * 2;
-  const innerSize = innerRadius * 2;
-  const outerSize = outerRadius * 2;
-  const outgoingDominant = outerRadius > innerRadius;
-  const incomingDominant = innerRadius > outerRadius;
-  return (
-    <div className="scale-legend-row">
-      <span
-        className={`scale-legend-circle-example ${
-          incomingDominant ? 'scale-legend-circle-example-incoming' : ''
-        }`}
-        style={{width: size, height: size}}
-      >
-        <span
-          className="scale-legend-circle-outer"
-          style={{
-            width: outgoingDominant ? outerSize : innerSize,
-            height: outgoingDominant ? outerSize : innerSize,
-            backgroundColor: rgbaToCss(
-              outgoingDominant ? colors.outgoing : colors.incoming,
-            ),
-            borderColor: outgoingDominant
-              ? rgbaToCss(mixRgba(colors.incoming, colors.outgoing, 0.4))
-              : rgbaToCss(colors.incoming),
-          }}
-        />
-        {incomingDominant ? (
-          <span
-            className="scale-legend-circle-ring"
-            style={{
-              width: outerSize,
-              height: outerSize,
-              borderColor: rgbaToCss(colors.outgoing),
-            }}
-          />
-        ) : (
-          <span
-            className="scale-legend-circle-inner"
-            style={{
-              width: innerSize,
-              height: innerSize,
-              backgroundColor: rgbaToCss(colors.incoming),
-            }}
-          />
-        )}
-      </span>
-      <span>{label}</span>
-    </div>
-  );
-}
-
-function formatLegendRange([min, max]: [number, number]): string {
-  return `${formatLegendValue(min)} - ${formatLegendValue(max)}`;
-}
-
-function formatLegendValue(value: number): string {
-  return value >= 1000
-    ? value.toLocaleString(undefined, {maximumFractionDigits: 0})
-    : value.toLocaleString(undefined, {maximumFractionDigits: 2});
-}
-
-function rgbaToCss([r, g, b, a]: [number, number, number, number]): string {
-  return `rgba(${r}, ${g}, ${b}, ${a / 255})`;
-}
-
-function mixRgba(
-  from: [number, number, number, number],
-  to: [number, number, number, number],
-  amount: number,
-): [number, number, number, number] {
-  return [
-    Math.round(from[0] * (1 - amount) + to[0] * amount),
-    Math.round(from[1] * (1 - amount) + to[1] * amount),
-    Math.round(from[2] * (1 - amount) + to[2] * amount),
-    Math.round(from[3] * (1 - amount) + to[3] * amount),
-  ];
 }
 
 function getTooltipState(
